pyproject.toml
[tool.poetry]
name = "image-processing"
version = "0.1.0"
description = ""
authors = ["Maximilian Graf"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
#python = "^3.12"
python = "<3.14,>=3.10"
kagglehub = "^0.3.13"
torch = "^2.8.0"
pillow = "^12.0.0"
matplotlib = "^3.10.7"
torchvision = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
